refactor(admin): extract Settings > Company info into standalone component
Behaviour-preserving Tier-0 decomposition of the 19k-line admin.component.ts.
Move the company-info panel (template + companyForm state + loadCompany/
companyMissingFields/saveCompany) into a new standalone OnPush component
(app-admin-company-info) under pages/admin/settings/. The shared CMS
content-version bookkeeping stays on AdminComponent and is threaded into the
child via four stable bound callbacks (rememberContentVersion,
withExpectedVersion, handleContentConflict, forgetContentVersion) so all CMS
panels keep sharing one contentVersions map. The parent template now renders
<app-admin-company-info *ngIf="section()==='settings'">; the child self-loads
on init.

Tests: company behaviour/branch coverage moved out of admin.component.spec,
admin.component.branchfill.spec and admin.component.coverage.spec into the new
admin-company-info.component.spec; loadCompany dropped from the settings-load
spy lists. Build clean; 829 targeted specs green (4 netflow + admin.component +
branchfill + coverage + new spec); settings section screenshot matches baseline.
